test: move unresolved ref tests into validate_test.go
